Lots of changes..... 1. Move the dynamic templates and dcp page code to a new dynamic library libdinter.(so/a) 2. Resolved warnings during compilation 3. New template generator mechanism which gives better templates and more flexibility in mixing code/html 4. Moved afc properties to application.xml, ajax controllers to be configured in main config file 5. New classes Bigint and Bigdecimal for arithmetic operations on big numbers 6. Improved database orm layer with new approach to querying using the CibernateQuery object, chaining available 7. Better approach to pagination, adding parameters and result ordering etc available 8. Improved the Date and DateFormat classes to map Date objects to database dates 9. Add cleanUp method to Tasks for auto cleanup 10. Use the Server - listener method in CHSServer 11. Change File Format from PC to Unix for several files 12. Updated HTML documentation

+++ ffead-server/include/AMEFEncoder.h
@@ -631,22 +631,22 @@ public:
 				type = AMEFObject::BIG_INT_TYPE;
 				ind = 3;
 			}
-			else if(lon<4294967296ULL)
+			else if(lon<(long)4294967296ULL)
 			{
 				type = AMEFObject::INT_TYPE;
 				ind = 4;
 			}
-			else if(lon<1099511627776ULL)
+			else if(lon<(long)1099511627776ULL)
 			{
 				type = AMEFObject::VS_LONG_INT_TYPE;
 				ind = 5;
 			}
-			else if(lon<281474976710656ULL)
+			else if(lon<(long)281474976710656ULL)
 			{
 				type = AMEFObject::S_LONG_INT_TYPE;
 				ind = 6;
 			}
-			else if(lon<72057594037927936ULL)
+			else if(lon<(long)72057594037927936ULL)
 			{
 				type = AMEFObject::B_LONG_INT_TYPE;
 				ind = 7;
